converting more receiving items to ts
